+++ packages/mel-components/src/mel-select/index.ts
@@ -1,0 +1,5 @@
+import MelSelect from './src/index.vue';
+
+export {
+  MelSelect,
+};